pyproject.toml
[project]
name = "meridian"
version = "0.1.0"
description = "Meridian Autoreport — editorial annual-report generation pipeline."
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "jsonschema>=4.23",
    "rich>=13.7",
]

[tool.uv]
package = false
